feat: 메인페이지 변경 사항 반영
