+++ frontend/src/components/common/GitHubLoginButton/index.tsx
@@ -5,8 +5,8 @@ import styles from './styles';
 const GitHubLoginButton = () => {
   return (
     <a href={process.env.REACT_APP_GITHUB_LOGIN_URL} css={styles.wrapper}>
-      <GoMarkGithub size={24} />
-      <span css={styles.text}>GitHub</span>
+      <GoMarkGithub size={32} />
+      <p css={styles.text}>GitHub 로그인</p>
     </a>
   );
 };
